Undated in-code test function to have no side-effects (restores globals)
Added handeling of allocating a constant (level-0) unit clause in the resolution graph (via allocConstUnitResol).

Added set content printing in Set.h

Added PoEC verifycation capabilities via verifyPoEC() in ProofRebuilder.cc

In RebuildProof - Part I - added handeling in the case of finding a single Unit used as a backbone.
NOTE: if AnalyzeFinal  returns a cluase that is non-ic then it's allocation will need to be deferred to the end of the alogirithm. This is handled partialy now but will be more thoroughly implemented int the next update.

Solver now reports different types of Unsat results (unsat at level 0, unsat without res proof).

If a Resol node is found to have out-of-order parents (with regards to the resolution ordering), caused by the solver's simplify call, and if it is found that the new ordering requires more memory allocated (due to it having only ic parents originally, which changed as some parents were kept as non-ic instead of being removed), then the re-orders clause will be allocated at the end of the ResolAllocator to accomodate the needed space.
AS A RESULT - ResolAllocator Shrink will be disable for now - until it can be re-worked to not shrink in-place and in uid order, but instead copy the way the ClauseAllocator does garbage collection.